requirements.txt
PyQt5==5.15.11
numpy==2.3.2
matplotlib==3.10.5
pillow==11.3.0
sounddevice
